Refactor: "FireStore 위치 컴포넌트 폴더에서 src 폴더로 이동"
